chore: setup multi process activity in example app
