+++ examples/sdk-app-example/app/proguard-rules.pro
