+++ examples/sdk-app-example/app/src/main/AndroidManifest.xml
@@ -26,6 +26,12 @@
             </intent-filter>
         </activity>
 
+        <activity
+            android:name=".MultiProcessActivity"
+            android:label="@string/app_name"
+            android:process="com.example.bugsnag.android.diffprocess">
+        </activity>
+
         <!--
           Uses a manifest placeholder to specify the API key. This means that different keys
           can be supplied depending on the build variant. See
