+++ examples/sdk-app-example/app/src/main/java/com/example/bugsnag/android/MultiProcessActivity.kt
@@ -1,0 +1,11 @@
+package com.example.bugsnag.android
+
+import android.os.Bundle
+import android.view.View
+
+class MultiProcessActivity : BaseCrashyActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        findViewById<View>(R.id.multi_process_title).visibility = View.VISIBLE
+    }
+}
