+++ examples/sdk-app-example/app/src/main/res/values/strings.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+  <string name="multi_process_title">Activity started in different process</string>
   <string name="unhandled_exceptions">Unhandled Exceptions</string>
   <string name="trigger_fatal_crash">Trigger a fatal crash</string>
   <string name="trigger_native_crash">Trigger a native crash</string>
@@ -18,6 +19,10 @@
 
   <string name="documentation">Documentation</string>
   <string name="read_the_docs">Read the Docs</string>
+
+  <string name="multi_process">Multi process apps</string>
+  <string name="multi_process_start">Start activity in another process</string>
+
   <string name="app_name">Bugsnag</string>
-    <string name="trigger_err_msg">Triggered an error, check your dashboard!</string>
+  <string name="trigger_err_msg">Triggered an error, check your dashboard!</string>
 </resources>
